requirements.txt
openai~=1.61.1
python-dotenv~=1.0.1
sounddevice~=0.5.1
soundfile~=0.13.1
numpy~=2.0.2
pvporcupine~=3.0.5